AskAmma/requirements.txt
accelerate
python-dotenv
fastapi
uvicorn
transformers
torch
langchain